appCollector/requirements.txt
lywsd03mmc
xiaomi-mi-scale
paho-mqtt
azure-iot-hub
msrest
azure-iot-device
